+++ examples/complex/packs/kitchen/boot.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 # Add services/ as an additional autoload directory
-BOXWERK_CONFIG[:autoload_dirs] << 'services'
+Boxwerk.package.autoloader.push_dir('services')