+++ test/boxwerk/integration/package_boot_test.rb
@@ -188,5 +188,161 @@ module Boxwerk
 
       assert_equal 'widget', root_box.eval('Widget.value')
     end
+
+    def test_boxwerk_package_available_in_boot
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+      File.write(
+        File.join(a_dir, 'lib', 'class_a.rb'),
+        "class ClassA\n  def self.value = 'from_a'\nend\n",
+      )
+      File.write(
+        File.join(a_dir, 'boot.rb'),
+        "BOOT_PKG = Boxwerk.package\n",
+      )
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      result = boot_system
+      a_box = result[:box_manager].boxes['packs/a']
+
+      assert_instance_of Boxwerk::PackageContext, a_box.eval('BOOT_PKG')
+    end
+
+    def test_boxwerk_package_name
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+      File.write(
+        File.join(a_dir, 'lib', 'class_a.rb'),
+        "class ClassA\nend\n",
+      )
+      File.write(
+        File.join(a_dir, 'boot.rb'),
+        "BOOT_PKG_NAME = Boxwerk.package.name\n",
+      )
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      result = boot_system
+      a_box = result[:box_manager].boxes['packs/a']
+
+      assert_equal 'packs/a', a_box.eval('BOOT_PKG_NAME')
+    end
+
+    def test_boxwerk_package_root
+      create_package(@tmpdir)
+      File.write(
+        File.join(@tmpdir, 'boot.rb'),
+        "BOOT_IS_ROOT = Boxwerk.package.root?\n",
+      )
+
+      result = boot_system
+      root_box = result[:box_manager].boxes['.']
+
+      assert root_box.eval('BOOT_IS_ROOT')
+    end
+
+    def test_boxwerk_package_config_frozen
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+      File.write(
+        File.join(a_dir, 'lib', 'class_a.rb'),
+        "class ClassA\nend\n",
+      )
+      File.write(
+        File.join(a_dir, 'boot.rb'),
+        "BOOT_CONFIG_FROZEN = Boxwerk.package.config.frozen?\n",
+      )
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      result = boot_system
+      a_box = result[:box_manager].boxes['packs/a']
+
+      assert a_box.eval('BOOT_CONFIG_FROZEN')
+    end
+
+    def test_boxwerk_package_autoloader_push_dir
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+
+      models_dir = File.join(a_dir, 'models')
+      FileUtils.mkdir_p(models_dir)
+      File.write(
+        File.join(models_dir, 'gadget.rb'),
+        "class Gadget\n  def self.value = 'gadget'\nend\n",
+      )
+
+      File.write(
+        File.join(a_dir, 'boot.rb'),
+        "Boxwerk.package.autoloader.push_dir('models')\n",
+      )
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      result = boot_system
+      a_box = result[:box_manager].boxes['packs/a']
+
+      assert_equal 'gadget', a_box.eval('Gadget.value')
+    end
+
+    def test_boxwerk_package_autoloader_collapse
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+
+      concerns_dir = File.join(a_dir, 'lib', 'concerns')
+      FileUtils.mkdir_p(concerns_dir)
+      File.write(
+        File.join(concerns_dir, 'sortable.rb'),
+        "module Sortable\n  def self.value = 'sortable'\nend\n",
+      )
+
+      File.write(
+        File.join(a_dir, 'boot.rb'),
+        "Boxwerk.package.autoloader.collapse('lib/concerns')\n",
+      )
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      result = boot_system
+      a_box = result[:box_manager].boxes['packs/a']
+
+      assert_equal 'sortable', a_box.eval('Sortable.value')
+    end
+
+    def test_boxwerk_package_constant_in_box
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+      File.write(
+        File.join(a_dir, 'lib', 'class_a.rb'),
+        "class ClassA\nend\n",
+      )
+      File.write(File.join(a_dir, 'boot.rb'), "\n")
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      result = boot_system
+      a_box = result[:box_manager].boxes['packs/a']
+
+      pkg = a_box.eval('BOXWERK_PACKAGE')
+      assert_instance_of Boxwerk::PackageContext, pkg
+      assert_equal 'packs/a', pkg.name
+    end
+
+    def test_boxwerk_package_nil_outside_boot
+      a_dir = create_package_dir('a')
+      create_package(a_dir)
+      File.write(
+        File.join(a_dir, 'lib', 'class_a.rb'),
+        "class ClassA\nend\n",
+      )
+      File.write(File.join(a_dir, 'boot.rb'), "\n")
+
+      create_package(@tmpdir, dependencies: ['packs/a'])
+
+      boot_system
+
+      assert_nil Boxwerk.package
+    end
   end
 end